requirements.txt
asgiref==3.7.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
Django==4.2.7
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
googletrans==3.0.0
gTTS==2.4.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
SpeechRecognition==3.10.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
yandex-translater==6.0
